requirements.txt
torch>=2.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
mlflow>=2.8.0
kafka-python>=2.0.2
requests>=2.31.0
